+++ src/InferenceSession.php
@@ -39,60 +39,62 @@ class InferenceSession
         self::env();
 
         // session options
-        $sessionOptions = $this->ffi->new('OrtSessionOptions*');
-        $this->checkStatus(($this->api->CreateSessionOptions)(\FFI::addr($sessionOptions)));
+        $sessionOptions = new Pointer($this->ffi->new('OrtSessionOptions*'));
+        $this->checkStatus(($this->api->CreateSessionOptions)(\FFI::addr($sessionOptions->ptr)));
+        $sessionOptions->free = $this->api->ReleaseSessionOptions;
+
         if ($enableCpuMemArena) {
-            $this->checkStatus(($this->api->EnableCpuMemArena)($sessionOptions));
+            $this->checkStatus(($this->api->EnableCpuMemArena)($sessionOptions->ptr));
         } else {
-            $this->checkStatus(($this->api->DisableCpuMemArena)($sessionOptions));
+            $this->checkStatus(($this->api->DisableCpuMemArena)($sessionOptions->ptr));
         }
         if ($enableMemPattern) {
-            $this->checkStatus(($this->api->EnableMemPattern)($sessionOptions));
+            $this->checkStatus(($this->api->EnableMemPattern)($sessionOptions->ptr));
         } else {
-            $this->checkStatus(($this->api->DisableMemPattern)($sessionOptions));
+            $this->checkStatus(($this->api->DisableMemPattern)($sessionOptions->ptr));
         }
         if ($enableProfiling) {
-            $this->checkStatus(($this->api->EnableProfiling)($sessionOptions, $this->ortString($profileFilePrefix ?? 'onnxruntime_profile_')));
+            $this->checkStatus(($this->api->EnableProfiling)($sessionOptions->ptr, $this->ortString($profileFilePrefix ?? 'onnxruntime_profile_')));
         } else {
-            $this->checkStatus(($this->api->DisableProfiling)($sessionOptions));
+            $this->checkStatus(($this->api->DisableProfiling)($sessionOptions->ptr));
         }
         if (!is_null($executionMode)) {
-            $this->checkStatus(($this->api->SetSessionExecutionMode)($sessionOptions, $executionMode->value));
+            $this->checkStatus(($this->api->SetSessionExecutionMode)($sessionOptions->ptr, $executionMode->value));
         }
         if (!is_null($freeDimensionOverridesByDenotation)) {
             foreach ($freeDimensionOverridesByDenotation as $k => $v) {
-                $this->checkStatus(($this->api->AddFreeDimensionOverride)($sessionOptions, $k, $v));
+                $this->checkStatus(($this->api->AddFreeDimensionOverride)($sessionOptions->ptr, $k, $v));
             }
         }
         if (!is_null($freeDimensionOverridesByName)) {
             foreach ($freeDimensionOverridesByName as $k => $v) {
-                $this->checkStatus(($this->api->AddFreeDimensionOverrideByName)($sessionOptions, $k, $v));
+                $this->checkStatus(($this->api->AddFreeDimensionOverrideByName)($sessionOptions->ptr, $k, $v));
             }
         }
         if (!is_null($graphOptimizationLevel)) {
-            $this->checkStatus(($this->api->SetSessionGraphOptimizationLevel)($sessionOptions, $graphOptimizationLevel->value));
+            $this->checkStatus(($this->api->SetSessionGraphOptimizationLevel)($sessionOptions->ptr, $graphOptimizationLevel->value));
         }
         if (!is_null($interOpNumThreads)) {
-            $this->checkStatus(($this->api->SetInterOpNumThreads)($sessionOptions, $interOpNumThreads));
+            $this->checkStatus(($this->api->SetInterOpNumThreads)($sessionOptions->ptr, $interOpNumThreads));
         }
         if (!is_null($intraOpNumThreads)) {
-            $this->checkStatus(($this->api->SetIntraOpNumThreads)($sessionOptions, $intraOpNumThreads));
+            $this->checkStatus(($this->api->SetIntraOpNumThreads)($sessionOptions->ptr, $intraOpNumThreads));
         }
         if (!is_null($logSeverityLevel)) {
-            $this->checkStatus(($this->api->SetSessionLogSeverityLevel)($sessionOptions, $logSeverityLevel));
+            $this->checkStatus(($this->api->SetSessionLogSeverityLevel)($sessionOptions->ptr, $logSeverityLevel));
         }
         if (!is_null($logVerbosityLevel)) {
-            $this->checkStatus(($this->api->SetSessionLogVerbosityLevel)($sessionOptions, $logVerbosityLevel));
+            $this->checkStatus(($this->api->SetSessionLogVerbosityLevel)($sessionOptions->ptr, $logVerbosityLevel));
         }
         if (!is_null($logid)) {
-            $this->checkStatus(($this->api->SetSessionLogId)($sessionOptions, $logid));
+            $this->checkStatus(($this->api->SetSessionLogId)($sessionOptions->ptr, $logid));
         }
         if (!is_null($optimizedModelFilepath)) {
-            $this->checkStatus(($this->api->SetOptimizedModelFilePath)($sessionOptions, $this->ortString($optimizedModelFilepath)));
+            $this->checkStatus(($this->api->SetOptimizedModelFilePath)($sessionOptions->ptr, $this->ortString($optimizedModelFilepath)));
         }
         if (!is_null($sessionConfigEntries)) {
             foreach ($sessionConfigEntries as $k => $v) {
-                $this->checkStatus(($this->api->AddSessionConfigEntry)($sessionOptions, $k, $v));
+                $this->checkStatus(($this->api->AddSessionConfigEntry)($sessionOptions->ptr, $k, $v));
             }
         }
         foreach ($providers as $provider) {
@@ -102,13 +104,13 @@ class InferenceSession
             }
 
             if ($provider == 'CUDAExecutionProvider') {
-                $cudaOptions = $this->ffi->new('OrtCUDAProviderOptionsV2*');
-                $this->checkStatus(($this->api->CreateCUDAProviderOptions)(\FFI::addr($cudaOptions)));
-                $this->checkStatus(($this->api->SessionOptionsAppendExecutionProvider_CUDA_V2)($sessionOptions, $cudaOptions));
-                ($this->api->ReleaseCUDAProviderOptions)($cudaOptions);
+                $cudaOptions = new Pointer($this->ffi->new('OrtCUDAProviderOptionsV2*'));
+                $this->checkStatus(($this->api->CreateCUDAProviderOptions)(\FFI::addr($cudaOptions->ptr)));
+                $cudaOptions->free = $this->api->ReleaseCUDAProviderOptions;
+                $this->checkStatus(($this->api->SessionOptionsAppendExecutionProvider_CUDA_V2)($sessionOptions->ptr, $cudaOptions->ptr));
             } elseif ($provider == 'CoreMLExecutionProvider') {
                 $coremlFlags = 0;
-                $this->checkStatus($this->ffi->OrtSessionOptionsAppendExecutionProvider_CoreML($sessionOptions, $coremlFlags));
+                $this->checkStatus($this->ffi->OrtSessionOptionsAppendExecutionProvider_CoreML($sessionOptions->ptr, $coremlFlags));
             } elseif ($provider == 'CPUExecutionProvider') {
                 break;
             } else {
@@ -120,13 +122,6 @@ class InferenceSession
         $this->allocator = $this->loadAllocator();
         $this->inputs = $this->loadInputs();
         $this->outputs = $this->loadOutputs();
-
-        ($this->api->ReleaseSessionOptions)($sessionOptions);
-    }
-
-    public function __destruct()
-    {
-        ($this->api->ReleaseSession)($this->session);
     }
 
     public function run($outputNames, $inputFeed, $logSeverityLevel = null, $logVerbosityLevel = null, $logid = null, $terminate = null)
@@ -156,35 +151,33 @@ class InferenceSession
         $outputNodeNames = $this->createNodeNames($outputNames, $refs);
 
         // run options
-        $runOptions = $this->ffi->new('OrtRunOptions*');
-        $this->checkStatus(($this->api->CreateRunOptions)(\FFI::addr($runOptions)));
+        $runOptions = new Pointer($this->ffi->new('OrtRunOptions*'));
+        $this->checkStatus(($this->api->CreateRunOptions)(\FFI::addr($runOptions->ptr)));
+        $runOptions->free = $this->api->ReleaseRunOptions;
+
         if (!is_null($logVerbosityLevel)) {
-            $this->checkStatus(($this->api->RunOptionsSetRunLogSeverityLevel)($runOptions, $logSeverityLevel));
+            $this->checkStatus(($this->api->RunOptionsSetRunLogSeverityLevel)($runOptions->ptr, $logSeverityLevel));
         }
         if (!is_null($logVerbosityLevel)) {
-            $this->checkStatus(($this->api->RunOptionsSetRunLogVerbosityLevel)($runOptions, $logVerbosityLevel));
+            $this->checkStatus(($this->api->RunOptionsSetRunLogVerbosityLevel)($runOptions->ptr, $logVerbosityLevel));
         }
         if (!is_null($logid)) {
-            $this->checkStatus(($this->api->RunOptionsSetRunTag)($runOptions, $logid));
+            $this->checkStatus(($this->api->RunOptionsSetRunTag)($runOptions->ptr, $logid));
         }
         if (!is_null($terminate)) {
             if ($terminate) {
-                $this->checkStatus(($this->api->RunOptionsSetTerminate)($runOptions));
+                $this->checkStatus(($this->api->RunOptionsSetTerminate)($runOptions->ptr));
             } else {
-                $this->checkStatus(($this->api->RunOptionsUnsetTerminate)($runOptions));
+                $this->checkStatus(($this->api->RunOptionsUnsetTerminate)($runOptions->ptr));
             }
         }
 
-        $this->checkStatus(($this->api->Run)($this->session, $runOptions, $inputNodeNames, $inputTensor, count($inputFeed), $outputNodeNames, count($outputNames), $outputTensor));
+        $this->checkStatus(($this->api->Run)($this->session->ptr, $runOptions->ptr, $inputNodeNames, $inputTensor, count($inputFeed), $outputNodeNames, count($outputNames), $outputTensor));
 
         $output = [];
         foreach ($outputTensor as $t) {
             $output[] = new OrtValue($t);
         }
-
-        // TODO use finally
-        ($this->api->ReleaseRunOptions)($runOptions);
-
         return $output;
     }
 
@@ -209,16 +202,17 @@ class InferenceSession
         $producerName = $this->ffi->new('char*');
         $version = $this->ffi->new('int64_t');
 
-        $metadata = $this->ffi->new('OrtModelMetadata*');
-        $this->checkStatus(($this->api->SessionGetModelMetadata)($this->session, \FFI::addr($metadata)));
+        $metadata = new Pointer($this->ffi->new('OrtModelMetadata*'));
+        $this->checkStatus(($this->api->SessionGetModelMetadata)($this->session->ptr, \FFI::addr($metadata->ptr)));
+        $metadata->free = $this->api->ReleaseModelMetadata;
 
         $customMetadataMap = [];
-        $this->checkStatus(($this->api->ModelMetadataGetCustomMetadataMapKeys)($metadata, $this->allocator, \FFI::addr($keys), \FFI::addr($numKeys)));
+        $this->checkStatus(($this->api->ModelMetadataGetCustomMetadataMapKeys)($metadata->ptr, $this->allocator->ptr, \FFI::addr($keys), \FFI::addr($numKeys)));
         for ($i = 0; $i < $numKeys->cdata; $i++) {
             $keyPtr = $keys[$i];
             $key = \FFI::string($keyPtr);
             $value = $this->ffi->new('char*');
-            $this->checkStatus(($this->api->ModelMetadataLookupCustomMetadataMap)($metadata, $this->allocator, $key, \FFI::addr($value)));
+            $this->checkStatus(($this->api->ModelMetadataLookupCustomMetadataMap)($metadata->ptr, $this->allocator->ptr, $key, \FFI::addr($value)));
             $customMetadataMap[$key] = \FFI::string($value);
 
             $this->allocatorFree($keyPtr);
@@ -226,12 +220,12 @@ class InferenceSession
         }
         $this->allocatorFree($keys);
 
-        $this->checkStatus(($this->api->ModelMetadataGetDescription)($metadata, $this->allocator, \FFI::addr($description)));
-        $this->checkStatus(($this->api->ModelMetadataGetDomain)($metadata, $this->allocator, \FFI::addr($domain)));
-        $this->checkStatus(($this->api->ModelMetadataGetGraphName)($metadata, $this->allocator, \FFI::addr($graphName)));
-        $this->checkStatus(($this->api->ModelMetadataGetGraphDescription)($metadata, $this->allocator, \FFI::addr($graphDescription)));
-        $this->checkStatus(($this->api->ModelMetadataGetProducerName)($metadata, $this->allocator, \FFI::addr($producerName)));
-        $this->checkStatus(($this->api->ModelMetadataGetVersion)($metadata, \FFI::addr($version)));
+        $this->checkStatus(($this->api->ModelMetadataGetDescription)($metadata->ptr, $this->allocator->ptr, \FFI::addr($description)));
+        $this->checkStatus(($this->api->ModelMetadataGetDomain)($metadata->ptr, $this->allocator->ptr, \FFI::addr($domain)));
+        $this->checkStatus(($this->api->ModelMetadataGetGraphName)($metadata->ptr, $this->allocator->ptr, \FFI::addr($graphName)));
+        $this->checkStatus(($this->api->ModelMetadataGetGraphDescription)($metadata->ptr, $this->allocator->ptr, \FFI::addr($graphDescription)));
+        $this->checkStatus(($this->api->ModelMetadataGetProducerName)($metadata->ptr, $this->allocator->ptr, \FFI::addr($producerName)));
+        $this->checkStatus(($this->api->ModelMetadataGetVersion)($metadata->ptr, \FFI::addr($version)));
 
         $ret = [
             'custom_metadata_map' => $customMetadataMap,
@@ -244,7 +238,6 @@ class InferenceSession
         ];
 
         // TODO use finally
-        ($this->api->ReleaseModelMetadata)($metadata);
         $this->allocatorFree($description);
         $this->allocatorFree($domain);
         $this->allocatorFree($graphName);
@@ -258,8 +251,12 @@ class InferenceSession
     public function endProfiling()
     {
         $out = $this->ffi->new('char*');
-        $this->checkStatus(($this->api->SessionEndProfiling)($this->session, $this->allocator, \FFI::addr($out)));
-        return \FFI::string($out);
+        $this->checkStatus(($this->api->SessionEndProfiling)($this->session->ptr, $this->allocator->ptr, \FFI::addr($out)));
+        try {
+            return \FFI::string($out);
+        } finally {
+            $this->allocatorFree($out);
+        }
     }
 
     // no way to set providers with C API yet
@@ -280,13 +277,14 @@ class InferenceSession
 
     private function loadSession($path, $sessionOptions)
     {
-        $session = $this->ffi->new('OrtSession*');
+        $session = new Pointer($this->ffi->new('OrtSession*'));
         if (is_resource($path) && get_resource_type($path) == 'stream') {
             $contents = stream_get_contents($path);
-            $this->checkStatus(($this->api->CreateSessionFromArray)(self::env(), $contents, strlen($contents), $sessionOptions, \FFI::addr($session)));
+            $this->checkStatus(($this->api->CreateSessionFromArray)(self::env()->ptr, $contents, strlen($contents), $sessionOptions->ptr, \FFI::addr($session->ptr)));
         } else {
-            $this->checkStatus(($this->api->CreateSession)(self::env(), $this->ortString($path), $sessionOptions, \FFI::addr($session)));
+            $this->checkStatus(($this->api->CreateSession)(self::env()->ptr, $this->ortString($path), $sessionOptions->ptr, \FFI::addr($session->ptr)));
         }
+        $session->free = $this->api->ReleaseSession;
         return $session;
     }
 
@@ -294,13 +292,15 @@ class InferenceSession
     {
         $inputs = [];
         $numInputNodes = $this->ffi->new('size_t');
-        $this->checkStatus(($this->api->SessionGetInputCount)($this->session, \FFI::addr($numInputNodes)));
+        $this->checkStatus(($this->api->SessionGetInputCount)($this->session->ptr, \FFI::addr($numInputNodes)));
         for ($i = 0; $i < $numInputNodes->cdata; $i++) {
             $namePtr = $this->ffi->new('char*');
-            $this->checkStatus(($this->api->SessionGetInputName)($this->session, $i, $this->allocator, \FFI::addr($namePtr)));
-            // freed in nodeInfo
-            $typeinfo = $this->ffi->new('OrtTypeInfo*');
-            $this->checkStatus(($this->api->SessionGetInputTypeInfo)($this->session, $i, \FFI::addr($typeinfo)));
+            $this->checkStatus(($this->api->SessionGetInputName)($this->session->ptr, $i, $this->allocator->ptr, \FFI::addr($namePtr)));
+
+            $typeinfo = new Pointer($this->ffi->new('OrtTypeInfo*'));
+            $this->checkStatus(($this->api->SessionGetInputTypeInfo)($this->session->ptr, $i, \FFI::addr($typeinfo->ptr)));
+            $typeinfo->free = $this->api->ReleaseTypeInfo;
+
             $inputs[] = array_merge(['name' => \FFI::string($namePtr)], $this->nodeInfo($typeinfo));
             $this->allocatorFree($namePtr);
         }
@@ -311,13 +311,15 @@ class InferenceSession
     {
         $outputs = [];
         $numOutputNodes = $this->ffi->new('size_t');
-        $this->checkStatus(($this->api->SessionGetOutputCount)($this->session, \FFI::addr($numOutputNodes)));
+        $this->checkStatus(($this->api->SessionGetOutputCount)($this->session->ptr, \FFI::addr($numOutputNodes)));
         for ($i = 0; $i < $numOutputNodes->cdata; $i++) {
             $namePtr = $this->ffi->new('char*');
-            $this->checkStatus(($this->api->SessionGetOutputName)($this->session, $i, $this->allocator, \FFI::addr($namePtr)));
-            // freed in nodeInfo
-            $typeinfo = $this->ffi->new('OrtTypeInfo*');
-            $this->checkStatus(($this->api->SessionGetOutputTypeInfo)($this->session, $i, \FFI::addr($typeinfo)));
+            $this->checkStatus(($this->api->SessionGetOutputName)($this->session->ptr, $i, $this->allocator->ptr, \FFI::addr($namePtr)));
+
+            $typeinfo = new Pointer($this->ffi->new('OrtTypeInfo*'));
+            $this->checkStatus(($this->api->SessionGetOutputTypeInfo)($this->session->ptr, $i, \FFI::addr($typeinfo->ptr)));
+            $typeinfo->free = $this->api->ReleaseTypeInfo;
+
             $outputs[] = array_merge(['name' => \FFI::string($namePtr)], $this->nodeInfo($typeinfo));
             $this->allocatorFree($namePtr);
         }
@@ -368,7 +370,7 @@ class InferenceSession
 
     private function allocatorFree($ptr)
     {
-        ($this->api->AllocatorFree)($this->allocator, $ptr);
+        ($this->api->AllocatorFree)($this->allocator->ptr, $ptr);
     }
 
     // wide string on Windows
@@ -400,12 +402,12 @@ class InferenceSession
         // TODO use mutex for thread-safety
 
         if (!isset(self::$env)) {
-            $env = FFI::instance()->new('OrtEnv*');
-            (self::api()->CreateEnv)(3, 'Default', \FFI::addr($env));
-            register_shutdown_function(fn () => (self::api()->ReleaseEnv)($env));
+            $env = new Pointer(FFI::instance()->new('OrtEnv*'));
+            (self::api()->CreateEnv)(3, 'Default', \FFI::addr($env->ptr));
+            $env->free = self::api()->ReleaseEnv;
             // disable telemetry
             // https://github.com/microsoft/onnxruntime/blob/master/docs/Privacy.md
-            self::checkStatus((self::api()->DisableTelemetryEvents)($env));
+            self::checkStatus((self::api()->DisableTelemetryEvents)($env->ptr));
             self::$env = $env;
         }