+++ src/Pointer.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace OnnxRuntime;
+
+class Pointer
+{
+    public $ptr;
+    public $free;
+
+    public function __construct($ptr, $free = null)
+    {
+        $this->ptr = $ptr;
+        $this->free = $free;
+    }
+
+    public function __destruct()
+    {
+        if (!is_null($this->free)) {
+            ($this->free)($this->ptr);
+        }
+    }
+}